Split ako into three packages
"ako" is a meta package, it has the same API as before
"ako_core" is the core package, it is not allowed to have dependencies other than flutter
"ako_optional" is a extension for the optional package, it depends on ako_core and optional

In the future we will have more packages which have additional dependencies.
Having all of them in one package would make it way to bloated.